Install scripts, build app with rollup & configurable mirror app id
