+++ README.md
@@ -30,13 +30,8 @@ Only tested on Linux. mDNS library issue to be fixed. `mdns` only works on Windo
 git clone https://github.com/hensm/caster.git
 npm install
 npm run build
+npm run install-manifest
 ````
-
-Installer scripts aren't written yet, so registering the native messaging manifest with Firefox and specifiying the path must be done manually:  
-[MDN: Native Manifests](https://developer.mozilla.org/en-US/Add-ons/WebExtensions/Native_manifests)
-
-`path` key within `app/caster_bridge.json` must be set to absolute path of `app/src/launcher.sh` or `app/src/launcher.bat`. Then, the manifest must be either moved to the correct location or the path added to the registry (Windows):  
-[MDN: Native Manifests # Manifest location](https://developer.mozilla.org/en-US/Add-ons/WebExtensions/Native_manifests#Manifest_location)
 
 ### Testing
