+++ app/.babelrc
@@ -1,0 +1,18 @@
+{
+    "presets": [
+        [
+            "@babel/preset-env"
+          , {
+                "targets": {
+                    "node": "current"
+                }
+            }
+        ]
+    ],
+    "plugins": [
+        "@babel/plugin-syntax-dynamic-import"
+      , "@babel/plugin-syntax-import-meta"
+      , "@babel/plugin-proposal-class-properties"
+      , "@babel/plugin-proposal-json-strings"
+    ]
+}
