+++ app/package.json
@@ -1,6 +1,38 @@
 {
-	"dependencies": {
-		"castv2": "^0.1.9",
-		"mdns-js": "^1.0.1"
-	}
+        "name": "fx_cast_bridge",
+        "license": "MIT",
+        "version": "0.1.0-alpha1",
+        "scripts": {
+                "build": "node scripts/build.js",
+                "install-manifest": "node scripts/install-manifest.js",
+                "start-debug": "node scripts/start-debug.js",
+                "start": "node dist/app.js"
+        },
+        "dependencies": {
+                "@babel/runtime": "^7.0.0",
+                "castv2": "^0.1.9",
+                "mdns-js": "^1.0.1",
+                "mkdirp": "^0.5.1"
+        },
+        "devDependencies": {
+                "@babel/core": "^7.0.0",
+                "@babel/plugin-proposal-class-properties": "^7.0.0",
+                "@babel/plugin-proposal-json-strings": "^7.0.0",
+                "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+                "@babel/plugin-syntax-import-meta": "^7.0.0",
+                "@babel/plugin-transform-runtime": "^7.0.0",
+                "@babel/preset-env": "^7.0.0",
+                "@babel/register": "^7.0.0",
+                "magic-string": "^0.25.0",
+                "minimist": "^1.2.0",
+                "rollup": "^0.66.0",
+                "rollup-plugin-babel": "^4.0.3",
+                "rollup-plugin-commonjs": "^9.1.3",
+                "rollup-plugin-json": "^3.1.0",
+                "rollup-plugin-node-builtins": "^2.1.2",
+                "rollup-plugin-node-resolve": "^3.4.0"
+        },
+        "optionalDependencies": {
+                "regedit": "^3.0.0"
+        }
 }
