+++ app/src/launcher.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
-
-AVAHI_COMPAT_NOWARN=1 node main.js
-
+#!/usr/bin/env bash
+BASEDIR=$(dirname $0)
+PATH=/usr/local/bin:$PATH
+AVAHI_COMPAT_NOWARN=1 node $BASEDIR/app.js
