--- app/uninstall.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-# TODO
