+++ package.json
@@ -3,8 +3,9 @@
         "postinstall": "npm run postinstall:app && npm run postinstall:ext",
         "postinstall:app": "npm install ./app --prefix ./app",
         "postinstall:ext": "npm install ./ext --prefix ./ext",
-        "build": "npm run build --prefix ./ext",
-        "test": "jasmine --config=jasmine.json"
+        "build": "npm run build --prefix ./app && npm run build --prefix ./ext --",
+        "test": "jasmine --config=jasmine.json",
+        "install-manifest": "npm run install-manifest --prefix ./app"
     },
     "devDependencies": {
         "jasmine": "^3.1.0",
